backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file backend/requirements.txt
asgiref==3.8.1
    # via django
django==5.2.3
    # via
    #   django-filter
    #   djangorestframework
django-filter==25.1
    # via lucky (pyproject.toml)
djangorestframework==3.16.0
    # via lucky (pyproject.toml)
pillow==11.2.1
    # via lucky (pyproject.toml)
psycopg==3.2.9
    # via lucky (pyproject.toml)
python-dotenv==1.1.0
    # via lucky (pyproject.toml)
sqlparse==0.5.3
    # via django
